+++ src/processes/integrated/mod.rs
@@ -1,4 +1,3 @@
-pub mod fastq_converter;
 pub mod merge_sam_pairs;
 pub mod num_procs;
 pub mod preprocess;